refactor: nmvc 및 mvc 패키지 구조 통합
