+++ jwp-basic-gradle/src/main/java/core/mvcframework/mapping/annotation/ControllerScanner.java
@@ -1,4 +1,4 @@
-package core.nmvc;
+package core.mvcframework.mapping.annotation;
 
 import com.google.common.collect.Maps;
 import core.annotation.Controller;
