+++ jwp-basic-gradle/src/test/java/core/nmvc/ControllerScannerTest.java
@@ -2,15 +2,12 @@ package core.nmvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import core.annotation.Controller;
-import core.ref.Student;
+import core.mvcframework.mapping.annotation.ControllerScanner;
 import java.util.Map;
-import java.util.Set;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.reflections.Reflections;
 
 public class ControllerScannerTest {
